Update set of students for dashboard endpoint too, factor out dashboard code to DashboardQueries
