+++ app/controllers/schools_controller.rb
@@ -29,34 +29,15 @@ class SchoolsController < ApplicationController
   end
 
   def absence_dashboard_data
-    student_absence_data = active_students(@school)
-      .includes([homeroom: :educator], :dashboard_absences, :sst_notes)
-    student_absence_data_json = student_absence_data.map do |student|
-      individual_student_absence_data(student)
-    end
-
-    render json: student_absence_data_json
+    render json: dashboard_queries.absence_dashboard_data(@school)
   end
 
   def tardies_dashboard_data
-    student_tardies_data = active_students(@school)
-      .includes([homeroom: :educator], :dashboard_tardies, :sst_notes)
-    student_tardies_data_json = student_tardies_data.map do |student|
-      individual_student_tardies_data(student)
-    end
-
-    render json: student_tardies_data_json
+    render json: dashboard_queries.student_tardies_data(@school)
   end
 
   def discipline_dashboard_data
-    student_discipline_data = active_students(@school)
-      .includes([homeroom: :educator], :discipline_incidents, :sst_notes)
-      .where('occurred_at >= ?', 1.year.ago).references(:discipline_incidents)
-    student_discipline_data_json = student_discipline_data.map do |student|
-      individual_student_discipline_data(student)
-    end
-
-    render json: student_discipline_data_json
+    render json: dashboard_queries.student_discipline_data(@school)
   end
 
   # This endpoint is internal-only for now, because of the authorization complexity.
@@ -167,48 +148,7 @@ class SchoolsController < ApplicationController
     end
   end
 
-  # Methods for Dashboards #
-  def shared_student_fields(student)
-    student.as_json(only: [
-      :first_name,
-      :last_name,
-      :grade,
-      :id
-    ]).merge({
-      homeroom_label: homeroom_label(student.homeroom),
-      sst_notes: student.sst_notes.as_json(only: [:event_note_type_id, :recorded_at])
-    })
-  end
-
-  def individual_student_absence_data(student)
-    shared_student_fields(student).merge({
-      absences: student.dashboard_absences.as_json(only: [:student_id, :occurred_at])
-    })
-  end
-
-  def individual_student_tardies_data(student)
-    shared_student_fields(student).merge({
-      tardies: student.dashboard_tardies.as_json(only: [:student_id, :occurred_at])
-    })
-  end
-
-  def individual_student_discipline_data(student)
-    shared_student_fields(student).merge({
-      discipline_incidents: student.discipline_incidents.as_json(only: [
-        :student_id,
-        :incident_code,
-        :incident_location,
-        :has_exact_time,
-        :occurred_at
-      ])
-    })
-  end
-
-  def active_students(school)
-    school.students.active
-  end
-
-  def homeroom_label(homeroom)
-    homeroom.try(:educator).try(:full_name) || homeroom.try(:name) || "No Homeroom"
+  def dashboard_queries
+    DashboardQueries.new(current_educator)
   end
 end
